move Layout files to new folder and update imports
