+++ components/Layout/Layout.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
-import AppNav from './AppNav'
-import Footer from './Footer'
+import AppNav from '../AppNav'
+import Footer from '../Footer'
 import Head from 'next/head'
-import type { LayoutGetter } from '../@types/page'
+import type { LayoutGetter } from '../../@types/page'
 import styles from '../scss/layout.module.scss'
 
 const Layout: React.FC<{ title?: string }> = ({ children, title }) => (
