+++ components/Layout/index.tsx
@@ -1,0 +1,1 @@
+export { getLayout, default } from './Layout'
